Update and rename README.md to UltrasonicSensor_integratedwith_WS2812LED.py
This is a code written in Arduino IDE to control a WS2812 LED strip with an ultrasonic sensor HC-SR04. The code sets up the pins, initializes the FastLED library for the LED strip, and continuously loops over taking distance readings from the ultrasonic sensor.

If the distance measured is less than 10 cm, all the LEDs on the strip are set to red color. If the distance is between 10cm and 30cm, all the LEDs are set to blue color. Further, this LED color changes to green if distance measured by ultrasonic sensor is between 30cm and 50cm.